Successful database with MySQL and Liquibase
Reading documentation and iteratively trying things finally lead to success. The test user needs to be created outside of the program which is easily done with MySQL workbench. Another method that will be implemented in the program is a data seed service. Properties and controls to ensure that isn't run in prod - or just removing the seeder completely from the program before release - are a later priority.